Revert moving BarrageMessageProducer; a relatively server side thing
