+++ server/src/main/java/io/deephaven/server/auth/AuthContextModule.java
@@ -11,7 +11,7 @@ import io.deephaven.auth.BasicAuthMarshaller;
 import io.deephaven.configuration.Configuration;
 import io.deephaven.internal.log.LoggerFactory;
 import io.deephaven.io.logger.Logger;
-import io.deephaven.extensions.barrage.BarrageStreamGenerator;
+import io.deephaven.extensions.barrage.BarrageStreamGeneratorImpl;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -19,7 +19,7 @@ import java.util.Map;
 
 @Module
 public class AuthContextModule {
-    private static final Logger log = LoggerFactory.getLogger(BarrageStreamGenerator.class);
+    private static final Logger log = LoggerFactory.getLogger(BarrageStreamGeneratorImpl.class);
     private static final String[] AUTH_HANDLERS =
             Configuration.getInstance().getStringArrayFromProperty("AuthHandlers");
