+++ server/src/main/java/io/deephaven/server/util/Scheduler.java
@@ -1,7 +1,7 @@
 /**
  * Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
  */
-package io.deephaven.util;
+package io.deephaven.server.util;
 
 import io.deephaven.base.clock.Clock;
 import org.jetbrains.annotations.NotNull;
